restrict deletes to only admins or users who authored the posts

+++ rareapi/views/Comments.py
@@ -86,12 +86,22 @@ class CommentViewSet(ViewSet):
     def destroy(self, request, pk=None):
         try:
             comment = Comments.objects.get(pk=pk)
-            comment.delete()
-            return Response({}, status=status.HTTP_204_NO_CONTENT)
         except Comments.DoesNotExist as ex:
             return Response({'message': ex.args[0]})
+
+        rare_user = RareUsers.objects.get(user=request.auth.user)
+        if (comment.author != rare_user) and (not request.auth.user.is_staff):
+            return Response(
+                {'message': 'Comments can only be deleted by admins or the users who authored them.'},
+                status=status.HTTP_403_FORBIDDEN
+            )
+        
+        try:
+            comment.delete()
         except Exception as ex:
             return Response({'message': ex.args[0]}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
+
+        return Response({}, status=status.HTTP_204_NO_CONTENT)
